Include aerosol dry deposition for GOCART aerosols (#47)
Adds a dry deposition module for GOCART aerosols,
chem/module_aer_drydep_simple.F, that is used automatically when
aer_drydep_opt = 1 and a GOCART chem_opt is selected. Previously,
enabling aer_drydep_opt=1 was possible but there was no code for dry
deposition in GOCART.

The dry deposition module follows Emmerson et al. (2020), based on the
Zhang et al. (2001) model but with revised parameters, especially for
the interception term. This increases dry deposition in the accumulation
mode.

The new dry deposition module could be used with other aerosol schemes
than GOCART, as long as the correct parameters are provided in the init
routine, but the current structure assumes externally mixed aerosols.

This fixes issue #27.